refactor: divide di modules
